backend/requirements.txt
mcp
anthropic
python-dotenv
websockets
